requirements.txt
BeautifulSoup==3.2.1
Django==1.5.4
Markdown==2.3.1
Pillow
South==0.7.6
argparse==1.2.1
django-classy-tags==0.4
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-localeurl==1.5
django-mptt==0.5.2
django-xmlrpc==0.1.4
html5lib==0.95
pyparsing==1.5.7
pytz==2013b
six==1.3.0
sorl-thumbnail==11.12
wsgiref==0.1.2
django-tinymce==1.5.1
https://pypi.python.org/packages/source/D/Django-tinymce-filebrowser/Django-tinymce-filebrowser.tar.gz#egg=Django-tinymce-filebrowser
ipython==1.1.0
requests
arrow
python-memcached
django-pagination